split up, add script_component, minor config beauitfication
