+++ addons/loadingscreen/CfgFunctions.hpp
@@ -1,0 +1,8 @@
+class CfgFunctions {
+	class full_mission_load {
+		class functions {
+			file = "\full_mission_load\functions";
+			class load {};
+		};
+	};
+};
