+++ addons/loadingscreen/script_component.hpp
@@ -1,0 +1,5 @@
+#define COMPONENT loadingscreen
+#define COMPONENT_BEAUTIFIED Loadingscreen
+
+#include "\z\ttt\addons\main\script_mod.hpp"
+#include "\z\ttt\addons\main\script_macros.hpp"
